+++ HSTracker/Core/Settings.swift
@@ -362,6 +362,10 @@ final class Settings {
     static var showMercsPlayerHover: Bool
     @UserDefault(key: Settings.show_mercs_tasks, defaultValue: true)
     static var showMercsTasks: Bool
+    @UserDefault(key: Settings.show_mercs_opponent_abilities, defaultValue: true)
+    static var showMercsOpponentAbilities
+    @UserDefault(key: Settings.show_mercs_player_abilities, defaultValue: true)
+    static var showMercsPlayerAbilities
 
     // MARK: - Importing
     @UserDefault(key: Settings.import_dungeon_include_passives, defaultValue: true)
@@ -568,6 +572,8 @@ extension Settings {
     static let show_mercs_opponent_hover = "show_mercs_opponent_hover"
     static let show_mercs_player_hover = "show_mercs_player_hover"
     static let show_mercs_tasks = "show_mercs_tasks"
+    static let show_mercs_opponent_abilities = "show_mercs_opponent_abilities"
+    static let show_mercs_player_abilities = "show_mercs_player_abilities"
     
     // MARK: - HSReplay.net related preferences
     static let hsreplay_upload_token = "hsreplay_upload_token"