+++ HSTracker/UIs/Trackers/BoardOverlay.swift
@@ -10,6 +10,49 @@ import Foundation
 import TextAttributes
 import AppKit
 
+class MercAbilityData {
+    var entity: Entity?
+    var card: Card?
+    var active = false
+    var gameTurn = 0
+    var hasTiers = false
+    
+    init(entity: Entity?, card: Card?, active: Bool, gameTurn: Int, hasTiers: Bool) {
+        self.entity = entity
+        self.card = card
+        self.active = active
+        self.gameTurn = gameTurn
+        self.hasTiers = hasTiers
+    }
+}
+
+private func getMercAbilities(player: Player) -> [[MercAbilityData]] {
+    let game = AppDelegate.instance().coreManager.game
+
+    let result: [[MercAbilityData]] = player.board.filter { x in x.isMinion }.sorted { (a, b) in a.zonePosition < b.zonePosition }.compactMap { entity in
+        let actualAbilities = player.playerEntities.filter { x in x[.lettuce_ability_owner] == entity.id && !x.has(tag: .lettuce_is_equipment) && !x.has(tag: .dont_show_in_history) && x.hasCardId }
+        let staticAbilities = RemoteConfig.mercenaries?.first { x in x.art_variation_ids.contains(entity.cardId)}?.abilities ?? [MercAbility]()
+        var data = [MercAbilityData]()
+        let max_ = min(3, max(staticAbilities.count, actualAbilities.count))
+        
+        for i in 0 ..< max_ {
+            let staticAbility = i < staticAbilities.count ? staticAbilities[i] : nil
+            let actual = staticAbility != nil ? actualAbilities.first(where: { x in staticAbility!.tier_ids.contains(x.cardId)}) : actualAbilities.first { x in data.all({ d in d.entity?.cardId != x.cardId }) }
+            if let actual = actual {
+                let active = entity[.lettuce_ability_tile_visual_self_only] == actual.id || entity[.lettuce_ability_tile_visual_all_visible] == actual.id
+                data.append(MercAbilityData(entity: actual, card: nil, active: active, gameTurn: 0, hasTiers: false))
+            } else if let staticAbility = staticAbility {
+                if let card = actual?.card ?? Cards.any(byId: staticAbility.tier_ids.last ?? "") {
+                    let gameTurn = game.gameEntity?[.turn] ?? 0
+                    data.append(MercAbilityData(entity: nil, card: card, active: false, gameTurn: gameTurn, hasTiers: staticAbility.tier_ids.count > 1))
+                }
+            }
+        }
+        return data
+    }
+    return result
+}
+
 class BoardMinionView: NSView {
     var entity: Entity?
     var card: Card?
@@ -71,9 +114,10 @@ class BoardMinionView: NSView {
             if game.isMercenariesMatch() && (game.gameEntity?[.step] == Step.main_combat.rawValue || !showMercenaryHover()) {
                 return
             }
-            let actualAbilities = AppDelegate.instance().coreManager.game.opponent.playerEntities.filter { x in x[.lettuce_ability_owner] == entity.id  && !entity.has(tag: .lettuce_is_equipment) && x.hasCardId && !x.has(tag: .dont_show_in_history)}
-            
-            let staticAbilities = RemoteConfig.mercenaries?.first { x in x.art_variation_ids.contains(entity.cardId)}?.abilities ?? [MercAbility]()
+            parent.updateAbilitiesVisibility(view: self)
+            let player: Player = playerType == .player ? game.player : game.opponent
+            let data = getMercAbilities(player: player)
+            let abilities = data[entity.zonePosition - 1]
             let hsFrame = SizeHelper.hearthstoneWindow.frame
             let h = hsFrame.height * 0.3
             let dh = hsFrame.height / 3.0
@@ -82,11 +126,10 @@ class BoardMinionView: NSView {
             let w = (CGFloat(nh) / CGFloat(388)) * 256.0
             let hoverFrame = NSRect(x: 0.0, y: 0.0, width: w, height: nh)
             let x = hsFrame.maxX - hoverFrame.width
-            let max_ = min(3, staticAbilities.count)
+            let max_ = min(3, abilities.count)
             for i in 0..<max_ {
-                let ability = staticAbilities[i]
-                let actual = actualAbilities.first { x in ability.tier_ids.contains(x.cardId) }
-                if let card = actual?.card ?? Cards.any(byId: ability.tier_ids.last ?? "") {
+                let abilityData = abilities[i]
+                if let card = abilityData.entity?.card ?? abilityData.card {
                     let frame = [x, hsFrame.maxY + delta - CGFloat(i+1) * h, hoverFrame.width, hoverFrame.height]
                     if i == 0 {
                         self.card = card
@@ -105,6 +148,7 @@ class BoardMinionView: NSView {
     
     override func mouseExited(with event: NSEvent) {
         self.parent.mousedOver = false
+        self.parent.clearAbilitiesVisibility()
         NotificationCenter.default
             .post(name: Notification.Name(rawValue: Events.hide_floating_card), object: nil,
                   userInfo: [ "card": self.card as Any])
@@ -114,37 +158,73 @@ class BoardMinionView: NSView {
 
 class BoardOverlayView: NSView {
     var minions: [BoardMinionView] = []
+    var abilities: [MercenariesAbilitiesView] = []
     var mousedOver: Bool = false
+    var updated: Date?
+    var playerType: PlayerType = .player
     
     init() {
         super.init(frame: NSRect.zero)
         
         addMinions()
     }
-
+    
     func updateBoardState(player: Player) {
-        let game = AppDelegate.instance().coreManager.game
-        let playerType = (player.id == game.player.id) ? PlayerType.player : PlayerType.opponent
-        if game.isMercenariesMatch() && mousedOver && game.gameEntity?[.step] == Step.main_combat.rawValue && minions.count > 0 {
-            NotificationCenter.default
-                .post(name: Notification.Name(rawValue: Events.hide_floating_card), object: nil,
-                      userInfo: [ "card": minions[0].card as Any])
-            mousedOver = false
+        if let updated = updated {
+            if updated.timeIntervalSinceNow > -0.05 {
+                return
+            }
         }
-        let oppBoard = player.board.filter({ x in x.isMinion }).sorted(by: { (a, b) in a[.zone_position] < b[.zone_position] })
-        let cnt = oppBoard.count
+        updated = Date()
+        let game = AppDelegate.instance().coreManager.game
+        let board = player.board.filter({ x in x.isMinion }).sorted(by: { (a, b) in a.zonePosition < b.zonePosition })
+
+        if game.isMercenariesMatch() && mousedOver && game.gameEntity?[.step] == Step.main_combat.rawValue {
+            if mousedOver {
+                NotificationCenter.default
+                    .post(name: Notification.Name(rawValue: Events.hide_floating_card), object: nil,
+                          userInfo: [ "card": minions[0].card as Any])
+                mousedOver = false
+            }
+            clearAbilitiesVisibility()
+            return
+        }
+        let cnt = board.count
+        if cnt == 0 || game.gameEnded {
+            isHidden = true
+        } else {
+            isHidden = false
+        }
         let frame = SizeHelper.opponentBoardOverlay()
         let w = SizeHelper.minionWidth
         let m = SizeHelper.mercenariesMinionMargin
+        let abilitySize = SizeHelper.abilitySize()
+        let overlayHeight = SizeHelper.boardOverlayHeight()
+        let yOff = playerType == .player ? abilitySize + overlayHeight * 0.14: 0
+        let aOff = playerType == .player ? 0 : overlayHeight + overlayHeight * 0.12
         let x = (frame.width - CGFloat(cnt) * w - 2.0 * CGFloat(cnt) * m) / 2.0
-        let rect = NSRect(x: x + m, y: 0, width: w, height: frame.height)
+        let rect = NSRect(x: x + m, y: yOff, width: w, height: overlayHeight)
+        let step = game.gameEntity?[.step] ?? 0
+        let showAbilities = game.isMercenariesMatch() && (step == Step.main_action.rawValue || step == Step.main_pre_action.rawValue)
+        let showPlayerAbilities = playerType == .player ? Settings.showMercsPlayerAbilities : Settings.showMercsOpponentAbilities
+        let mercAbilities = showAbilities && showPlayerAbilities ? getMercAbilities(player: player) : nil
         for i in 0..<cnt {
-            minions[i].frame = rect.offsetBy(dx: CGFloat(i) * (w + 2 * m), dy: 0)
-            minions[i].entity = oppBoard[i]
+            let fr = rect.offsetBy(dx: CGFloat(i) * (w + 2 * m), dy: 0)
+            minions[i].frame = fr
+            minions[i].entity = board[i]
             minions[i].playerType = playerType
             if !minions[i].isDescendant(of: self) {
                 addSubview(minions[i])
             }
+            
+            let abilityRect = NSRect(x: fr.minX, y: aOff, width: fr.width, height: abilitySize)
+            abilities[i].frame = abilityRect
+//            abilities[i].isHidden = playerType == .player ? !Settings.showMercsPlayerAbilities : !Settings.showMercsOpponentAbilities
+            if !abilities[i].isDescendant(of: self) {
+                addSubview(abilities[i])
+            }
+            
+            abilities[i].setAbilities(abilities: mercAbilities?[i], abilitySize: abilitySize, parentFrame: abilityRect)
         }
         for i in cnt..<7 {
             minions[i].frame = NSRect.zero
@@ -152,6 +232,10 @@ class BoardOverlayView: NSView {
             minions[i].playerType = playerType
             if minions[i].isDescendant(of: self) {
                 minions[i].removeFromSuperview()
+            }
+            abilities[i].frame = NSRect.zero
+            if abilities[i].isDescendant(of: self) {
+                abilities[i].removeFromSuperview()
             }
         }
         needsLayout = true
@@ -163,6 +247,48 @@ class BoardOverlayView: NSView {
             let minion = BoardMinionView(parent: self)
             minions.append(minion)
             addSubview(minion)
+            let ability = MercenariesAbilitiesView()
+            abilities.append(ability)
+            addSubview(ability)
+        }
+    }
+    
+    func updateAbilitiesVisibility(view: BoardMinionView) {
+        if let hoverIndex = minions.firstIndex(of: view) {
+            let game = AppDelegate.instance().coreManager.game
+            let player = playerType == .player ? game.player : game.opponent
+            let excludeIndex = playerType == .player ? game.gameEntity?.has(tag: .allow_move_minion) ?? false : false
+            let board = player?.board.filter({ x in x.isMinion }).sorted(by: { (a, b) in a.zonePosition < b.zonePosition }) ?? [Entity]()
+            let cnt = board.count
+
+            let center = (cnt + 1) / 2 - 1
+            
+            for i in 0 ..< 7 {
+                if hoverIndex <= center {
+                    // tooltip to the right
+                    if i <= hoverIndex {
+                        abilities[i].isHidden = false
+                    } else {
+                        abilities[i].isHidden = true
+                    }
+                } else {
+                    // tooltip to the left
+                    if i >= hoverIndex {
+                        abilities[i].isHidden = false
+                    } else {
+                        abilities[i].isHidden = true
+                    }
+                }
+            }
+            if excludeIndex {
+                abilities[hoverIndex].isHidden = true
+            }
+        }
+    }
+    
+    func clearAbilitiesVisibility() {
+        for ability in abilities {
+            ability.isHidden = false
         }
     }
 
@@ -197,5 +323,9 @@ class BoardOverlay: OverWindowController {
     
     func updateBoardState(player: Player) {
         view.updateBoardState(player: player)
+    }
+    
+    func setPlayerType(playerType: PlayerType) {
+        view.playerType = playerType
     }
 }